+++ winpr/libwinpr/crt/unicode.c
@@ -405,8 +405,8 @@ SSIZE_T ConvertWCharNToUtf8(const WCHAR* wstr, size_t wlen, char* str, size_t le
 		isNullTerminated = TRUE;
 		iwlen++;
 	}
-	const int rc =
-	    WideCharToMultiByte(CP_UTF8, 0, wstr, (int)iwlen, str, MIN(INT32_MAX, len), NULL, NULL);
+	const int rc = WideCharToMultiByte(CP_UTF8, 0, wstr, (int)iwlen, str, (int)MIN(INT32_MAX, len),
+	                                   NULL, NULL);
 	if ((rc <= 0) || ((len > 0) && ((size_t)rc > len)))
 		return -1;
 	else if (!isNullTerminated)
@@ -436,7 +436,7 @@ SSIZE_T ConvertMszWCharNToUtf8(const WCHAR* wstr, size_t wlen, char* str, size_t
 		return -1;
 	}
 
-	const int iwlen = MIN(INT32_MAX, len);
+	const int iwlen = (int)MIN(INT32_MAX, len);
 	const int rc = WideCharToMultiByte(CP_UTF8, 0, wstr, (int)wlen, str, iwlen, NULL, NULL);
 	if ((rc <= 0) || ((len > 0) && (rc > iwlen)))
 		return -1;
@@ -477,7 +477,7 @@ SSIZE_T ConvertUtf8NToWChar(const char* str, size_t len, WCHAR* wstr, size_t wle
 		ilen++;
 	}
 
-	const int iwlen = MIN(INT32_MAX, wlen);
+	const int iwlen = (int)MIN(INT32_MAX, wlen);
 	const int rc = MultiByteToWideChar(CP_UTF8, 0, str, (int)ilen, wstr, iwlen);
 	if ((rc <= 0) || ((wlen > 0) && (rc > iwlen)))
 		return -1;
@@ -508,7 +508,7 @@ SSIZE_T ConvertMszUtf8NToWChar(const char* str, size_t len, WCHAR* wstr, size_t 
 		return -1;
 	}
 
-	const int iwlen = MIN(INT32_MAX, wlen);
+	const int iwlen = (int)MIN(INT32_MAX, wlen);
 	const int rc = MultiByteToWideChar(CP_UTF8, 0, str, (int)len, wstr, iwlen);
 	if ((rc <= 0) || ((wlen > 0) && (rc > iwlen)))
 		return -1;
